drivers: stm32_rtc: add the capability to wakeup the platform
During probe, we look for the property "wakeup-source" that
will trigger the feature "RTC_WAKEUP_ALARM" which will
be send to the callers of the framework "rtc_get_info" PTA.

We also register the "stm32_rtc_alarm_wake_set_status" callback.
This callback should be called when the callers knows that the platform
will go in sleep mode.

Signed-off-by: Clément Le Goffic <clement.legoffic@foss.st.com>